Add error display and fix linter errors
