+++ src/assets/style.css
@@ -5,6 +5,14 @@ body {
   font-family: sans-serif;
 }
 
+.display-none {
+  display: none;
+}
+
+.display-block {
+  display: block;
+}
+
 .background-picture {
   width: 100%;
   height: 100%;
@@ -80,4 +88,21 @@ body {
 
 #submit-input:focus {
   outline: none;
+}
+
+#error {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  width: 300px;
+  height: 80px;
+  background-color: rgb(220, 20, 60);
+  border: 1px solid red;
+  text-align: center;
+  padding-top: 15px;
+}
+
+.error-message {
+  font-size: 30px;
 }
